==> analysis/discount/step0.ipynb <==
from pyspark.sql import SparkSession
from pyspark import SparkContext
import pyspark.sql.functions as func
import pandas as pd

sc = SparkContext(appName="discount analysis")
spark = SparkSession(sc)

input_df = spark.read.csv("input.txt")

# review - price history

import os
input_path = "../../data/review_price/review-download/output/"
input_list = os.listdir(input_path)
with open("input.txt", 'w', encoding='utf-8') as f:
    for file in input_list:
        f.write(file+'\n')

input_df = spark.read.csv("input.txt")
input_df = input_df.toPandas()
input_list = input_df[input_df.columns[0]].to_list()
print(type(input_list))
input_list[:10]

df = spark.read.csv(input_path+"240.csv", header=True, inferSchema=True)
df.show(10)

df.printSchema()

df = df.dropna()
df.show(10)

df.columns

df.select('discount').show(5)

df = df.withColumn("discount", func.round((df.Price - df.PriceChangeTo) / df.Price * 100, 0).cast('Int'))
df.show(10)

df = df.withColumn('app_id', func.lit('240'))
df.show()

# df.write.options(header='True', delimiter=',').csv("./ouput")

id1 = input_list[0]
df1 = spark.read.csv(input_path+id1, header=True, inferSchema=True)
df1 = df1.withColumn('app_id', func.lit(id1))
df1.show()

id2 = input_list[1]
df2 = spark.read.csv(input_path+id2, header=True, inferSchema=True)
df2 = df2.withColumn('app_id', func.lit(id2))
df2.show()

union_df = df1.union(df2)
union_df.show(100)
