==> analysis/discount/step1.ipynb <==
from pyspark.sql import SparkSession
from pyspark import SparkContext
import pyspark.sql.functions as func
from pyspark.sql import Window

sc = SparkContext(appName="step1")
spark = SparkSession(sc)

input_path = "./raw_input/"

df = spark.read.csv(input_path+"240.csv", header=True, inferSchema=True).dropna()
df = df.withColumn("discount", func.round((df.Price - df.PriceChangeTo) / df.Price * 100, 0).cast('Int'))
df = df.withColumn('app_id', func.lit('240'))
df.show(100)

df = df.withColumn("index", func.row_number().over(Window.orderBy(func.monotonically_increasing_id())))
df.show()

colums = ["date", "total_reviews", "total_positive", "total_negative", "review_score", "discount"]
for c in colums:
    df = df.withColumn("prev_"+c, func.lag(func.col(c), 1).over(Window.orderBy("index")))
for c in colums:
    df = df.withColumn("next_"+c, func.lead(func.col(c), 1).over(Window.orderBy("index")))
df.show(2)

res = df.filter(df.discount < 0).dropna()
res = res.withColumn("total_increase", res.total_reviews - res.prev_total_reviews)\
        .withColumn("positive_increase", res.total_positive - res.prev_total_positive)\
        .withColumn("negative_increase", res.total_negative - res.prev_total_negative)\
        .withColumn("days_increase", func.datediff(res.date, res.prev_date))\
        .withColumn("total_normal", res.next_total_reviews - res.total_reviews)\
        .withColumn("positive_normal", res.next_total_positive - res.total_positive)\
        .withColumn("negative_normal", res.next_total_negative - res.total_negative)\
        .withColumn("days_normal", func.datediff(res.next_date, res.date))\
        .withColumn("raw_price", res.PriceChangeTo)\
        .withColumn("sale_price", res.Price)\
        .withColumn("discount", res.prev_discount)#.\
        # select("date", "raw_price", "sale_price", "discount",\
        #         "total_increase", "positive_increase", "negative_increase", "days_increase",\
        #         "total_normal", "positive_normal", "negative_normal", "days_normal")

res = res.withColumn("total_increase_rate", res.total_increase / res.days_increase)\
        .withColumn("total_normal_rate", res.total_normal / res.days_normal)

res.select("date", "raw_price", "sale_price", "discount", "total_increase_rate", "total_normal_rate").show()

res.columns

res = res.drop('Price', 'PriceChangeTo')
res.columns

res = res.withColumn("sale_price_scale", (res.sale_price/10).cast('Int'))
res.select("sale_price_scale").show(5)

distinct_sale_price = res.dropDuplicates(['sale_price']).select("index", "sale_price").withColumn("historical_low", func.lit(1))
distinct_sale_price.show()

distinct_sale_price = distinct_sale_price.withColumn('prev_sale_price', \
    func.lag(func.col('sale_price'), 1).over(Window.orderBy("index")))
distinct_sale_price = distinct_sale_price.withColumn('diff', distinct_sale_price.sale_price - distinct_sale_price.prev_sale_price).fillna(-1)
distinct_sale_price.show()

distinct_sale_price.where(distinct_sale_price.diff > 0).count()

distinct_sale_price = distinct_sale_price.withColumn('historical_low', func.when(distinct_sale_price.diff > 0, 0).otherwise(1))\
    .where(distinct_sale_price.historical_low == 1)
distinct_sale_price.show()

distinct_sale_price = res.dropDuplicates(['sale_price']).select("index", "sale_price")

distinct_sale_price = distinct_sale_price.withColumn('prev_sale_price', \
    func.lag(func.col('sale_price'), 1).over(Window.orderBy("index")))

distinct_sale_price = distinct_sale_price.withColumn('diff', \
    distinct_sale_price.sale_price - distinct_sale_price.prev_sale_price).fillna(-1)

isRepeat = distinct_sale_price.where(distinct_sale_price.diff > 0).count() > 0

while isRepeat:
    distinct_sale_price = distinct_sale_price.withColumn('historical_low', \
        func.when(distinct_sale_price.diff > 0, 0).otherwise(1))
    distinct_sale_price = distinct_sale_price.where(distinct_sale_price.historical_low == 1)

    distinct_sale_price = distinct_sale_price.withColumn('prev_sale_price', \
        func.lag(func.col('sale_price'), 1).over(Window.orderBy("index")))
    
    distinct_sale_price = distinct_sale_price.withColumn('diff', \
        distinct_sale_price.sale_price - distinct_sale_price.prev_sale_price).fillna(-1)

    isRepeat = distinct_sale_price.where(distinct_sale_price.diff > 0).count() > 0

distinct_sale_price.show()

index_list = [row['index'] for row in distinct_sale_price.select('index').collect()]
index_list

res = res.withColumn('historical_low', func.when(res.index.isin(index_list), 1).otherwise(0))
res.select('sale_price', 'historical_low').show()

res.columns

tags_df = spark.read.csv("./tags_input/joint_category_genre.csv", header=True, inferSchema=True)
tags_df.show(5)

tags_df = tags_df.withColumn('category', func.split(func.col('category'), ',')).withColumn('genre', func.split(func.col('genre'), ','))
tags_df.show(5)

tags_df.printSchema()

joint_df = res.join(tags_df, res.app_id == tags_df.id, 'inner')
joint_df.show(5)

df1 = joint_df.select('app_id', 'index', func.explode('category').alias('id'))

df1.show(10)

test = df1.groupby('index').pivot('id').agg(func.lit(1))
test.show(100)

sc.stop()
