==> analysis/discount/analysis3.ipynb <==
from pyspark.sql import SparkSession
from pyspark import SparkContext
import pyspark.sql.functions as func
from pyspark.sql import Window

sc = SparkContext(appName="analysis2")
spark = SparkSession(sc)

df = spark.read.csv("./analysis_input/", header=True, inferSchema=True)
df.show(10)

df.groupBy('historical_low').agg(func.avg('sale_increase_rate'), func.avg('normal_increase_rate'), func.avg('effect_min'), func.avg('effect_plus')).show()

df.groupBy('historical_low').avg('sale_increase_rate', 'normal_increase_rate', 'effect_min', 'effect_plus').show()

df.groupBy('year').avg('sale_increase_rate', 'normal_increase_rate', 'effect_min', 'effect_plus').sort('year').show()

df.groupBy('review_score').avg('sale_increase_rate', 'normal_increase_rate', 'effect_min', 'effect_plus').sort('review_score').show()

df.groupBy('sale_price_scale').avg('sale_increase_rate', 'normal_increase_rate', 'effect_min', 'effect_plus').sort('sale_price_scale').show()

df.groupBy('sale_price_scale').avg('popularity').sort('sale_price_scale').show()

df.groupBy('discount').avg('sale_increase_rate', 'normal_increase_rate', 'effect_min', 'effect_plus').sort('discount').show()

sc.stop()
